Textract demo app now working, upload file, then view json results and download bounding box classified pdf
