+++ requirements.txt
@@ -1,9 +1,15 @@
+blinker==1.6.2
 boto3==1.26.138
 botocore==1.29.138
+click==8.1.3
 deprecation==2.1.0
+Flask==2.3.2
 img2pdf==0.4.4
+itsdangerous==2.1.2
+Jinja2==3.1.2
 jmespath==1.0.1
 lxml==4.9.2
+MarkupSafe==2.1.3
 numpy==1.24.3
 opencv-python==4.7.0.72
 opencv-python-headless==4.7.0.72
@@ -11,11 +17,13 @@ packaging==23.1
 pdf2image==1.16.3
 pikepdf==7.2.0
 Pillow==9.5.0
+pytesseract==0.3.10
 python-dateutil==2.8.2
 s3transfer==0.6.1
 six==1.16.0
 textract-trp==0.1.3
 urllib3==1.26.16
+Werkzeug==2.3.4
 aom
 aribb24
 brotli
